requirements.txt
numpy==2.2.6
pandas==2.2.3
matplotlib==3.10.3
seaborn==0.13.2
scikit-learn==1.6.1
lightgbm==4.6.0
optuna==4.4.0
shap==0.48.0
PyYAML==6.0.2
yamlcore==0.0.4